[feat] 게시글 조회 API 오류메시지 수정 및 Response 구조 변경 Jira: YP-285
