+++ yp-online/src/main/java/kr/co/yourplanet/online/business/instagram/controller/InstagramController.java
@@ -1,7 +1,7 @@
 package kr.co.yourplanet.online.business.instagram.controller;
 
 import kr.co.yourplanet.core.enums.StatusCode;
-import kr.co.yourplanet.online.business.instagram.dto.InstagramMediaForm;
+import kr.co.yourplanet.online.business.instagram.dto.InstagramMediasForm;
 import kr.co.yourplanet.online.business.instagram.service.InstagramService;
 import kr.co.yourplanet.online.common.ResponseForm;
 import kr.co.yourplanet.online.jwt.JwtPrincipal;
@@ -12,8 +12,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
-
 @RestController
 @RequiredArgsConstructor
 public class InstagramController {
@@ -21,15 +19,15 @@ public class InstagramController {
     private final InstagramService instagramService;
 
     @GetMapping("/instagram/medias")
-    public ResponseForm<List<InstagramMediaForm>> getMedias(@RequestParam(required = false) String permalink, @AuthenticationPrincipal JwtPrincipal principal) {
-        List<InstagramMediaForm> instagramMediaList;
+    public ResponseForm<InstagramMediasForm> getMedias(@RequestParam(required = false) String permalink, @AuthenticationPrincipal JwtPrincipal principal) {
+        InstagramMediasForm instagramMediasForm;
 
         if (StringUtils.hasText(permalink)) {
-            instagramMediaList = instagramService.getMediasByPermalink(permalink, principal.getId());
+            instagramMediasForm = instagramService.getMediasByPermalink(permalink, principal.getId());
         } else{
-            instagramMediaList = instagramService.getAllMediasByMemberId(principal.getId());
+            instagramMediasForm = instagramService.getAllMediasByMemberId(principal.getId());
         }
 
-        return new ResponseForm<>(StatusCode.OK, instagramMediaList);
+        return new ResponseForm<>(StatusCode.OK, instagramMediasForm);
     }
 }
