+++ yp-online/src/main/java/kr/co/yourplanet/online/business/instagram/service/InstagramService.java
@@ -1,11 +1,9 @@
 package kr.co.yourplanet.online.business.instagram.service;
 
-import kr.co.yourplanet.online.business.instagram.dto.InstagramMediaForm;
-
-import java.util.List;
+import kr.co.yourplanet.online.business.instagram.dto.InstagramMediasForm;
 
 public interface InstagramService {
 
-    List<InstagramMediaForm> getAllMediasByMemberId(Long memberId);
-    List<InstagramMediaForm> getMediasByPermalink(String permalink, Long memberId);
+    InstagramMediasForm getAllMediasByMemberId(Long memberId);
+    InstagramMediasForm getMediasByPermalink(String permalink, Long memberId);
 }
