+++ yp-online/src/main/java/kr/co/yourplanet/online/business/instagram/service/impl/InstagramServiceImpl.java
@@ -2,7 +2,7 @@ package kr.co.yourplanet.online.business.instagram.service.impl;
 
 import kr.co.yourplanet.core.entity.instagram.InstagramMedia;
 import kr.co.yourplanet.core.enums.StatusCode;
-import kr.co.yourplanet.online.business.instagram.dto.InstagramMediaForm;
+import kr.co.yourplanet.online.business.instagram.dto.InstagramMediasForm;
 import kr.co.yourplanet.online.business.instagram.repository.InstagramMediaRepository;
 import kr.co.yourplanet.online.business.instagram.service.InstagramService;
 import kr.co.yourplanet.online.common.exception.BusinessException;
@@ -14,7 +14,6 @@ import org.springframework.util.StringUtils;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
@@ -25,39 +24,39 @@ public class InstagramServiceImpl implements InstagramService {
     private static final String PERMALINK_PATTERN = ".*instagram.com/p/([^/]+).*";
 
     @Override
-    public List<InstagramMediaForm> getAllMediasByMemberId(Long memberId) {
+    public InstagramMediasForm getAllMediasByMemberId(Long memberId) {
 
         List<InstagramMedia> mediaList = instagramMediaRepository.findByMemberId(memberId);
         if (CollectionUtils.isEmpty(mediaList)) {
-            throw new BusinessException(StatusCode.NOT_FOUND, "등록된 인스타그램 게시물이 없습니다.", false);
+            throw new BusinessException(StatusCode.NOT_FOUND, "올바른 링크를 입력해 주세요.", false);
         }
 
-        return convertInstagramMediaEntityToDto(mediaList);
+        InstagramMediasForm instagramMediasForm = new InstagramMediasForm();
+        instagramMediasForm.setMediasByEntityList(mediaList);
+
+        return instagramMediasForm;
     }
 
     @Override
-    public List<InstagramMediaForm> getMediasByPermalink(String userInputPermalink, Long memberId) {
+    public InstagramMediasForm getMediasByPermalink(String userInputPermalink, Long memberId) {
 
         // userInputPermalink = https://www.instagram.com/p/unique-url/xxxxx
         String parsedPermalink = parsePermalink(userInputPermalink);
 
         if (!StringUtils.hasText(parsedPermalink)) {
-            throw new BusinessException(StatusCode.BAD_REQUEST, "유효하지 않은 형태의 URL입니다. : " + userInputPermalink, false);
+            throw new BusinessException(StatusCode.BAD_REQUEST, "올바른 링크를 입력해 주세요.", false);
         }
 
         List<InstagramMedia> mediaList = instagramMediaRepository.findByMemberIdAndPermalinkLike(memberId, parsedPermalink);
 
         if (CollectionUtils.isEmpty(mediaList)) {
-            throw new BusinessException(StatusCode.NOT_FOUND, "존재하지 인스타그램 URL입니다. : " + userInputPermalink, false);
+            throw new BusinessException(StatusCode.NOT_FOUND, "올바른 링크를 입력해 주세요.", false);
         }
 
-        for (InstagramMedia instagramMedia : mediaList) {
-            if (!memberId.equals(instagramMedia.getMemberId())) {
-                throw new BusinessException(StatusCode.FORBIDDEN, "자신의 인스타그램 게시글만 조회할 수 있습니다. : " + userInputPermalink, false);
-            }
-        }
+        InstagramMediasForm instagramMediasForm = new InstagramMediasForm();
+        instagramMediasForm.setMediasByEntityList(mediaList);
 
-        return convertInstagramMediaEntityToDto(mediaList);
+        return instagramMediasForm;
     }
 
     // 사용자가 입력한 인스타그램 URL에서 unique-url 파싱하는 함수
@@ -72,25 +71,6 @@ public class InstagramServiceImpl implements InstagramService {
         }
     }
 
-    private List<InstagramMediaForm> convertInstagramMediaEntityToDto(List<InstagramMedia> instagramMediaList) {
-        return instagramMediaList.stream()
-                .map(this::convertInstagramMediaEntityToDto)
-                .collect(Collectors.toList());
-    }
 
-    private InstagramMediaForm convertInstagramMediaEntityToDto(InstagramMedia instagramMedia) {
-        return InstagramMediaForm.builder()
-                .id(instagramMedia.getId())
-                .caption(instagramMedia.getCaption())
-                //.isSharedToFeed(true) 용도확인 필요
-                .mediaType(instagramMedia.getMediaType())
-                .mediaUrl(instagramMedia.getMediaUrl())
-                .permalink(instagramMedia.getPermalink())
-                .thumbnailUrl(instagramMedia.getPermalink() + "media/?size=l") // permalink 일정시간 지나면 expired 되어 대체하기 위함
-                //.timestamp() 배치프로그램 수정필요
-                .username(instagramMedia.getUsername())
-                .build();
-
-    }
 
 }
